// game/src/generated/worldMapFingerprint.ts
export const WORLD_MAP_IMAGE_BOUNDS = {"minX":-650,"maxX":1000,"minZ":-450,"maxZ":1200} as const;
export const WORLD_MAP_PLAYABLE_BOUNDS = {"minX":-560,"maxX":910,"minZ":-410,"maxZ":1150} as const;
export const WORLD_MAP_MINIMAP_RENDITION = {
  "id": "minimap",
  "role": "minimap",
  "path": "generated/world-map-minimap.webp",
  "format": "webp",
  "width": 800,
  "height": 800,
  "metresPerPixel": 2.0625,
  "bytes": 143116,
  "sha256": "268d90f2233330fde37aacab801898ae44ab5bfb4c00c9e93c34cf0e451659cb",
  "quality": 91
} as const;
export const WORLD_MAP_DETAIL_RENDITIONS = [
  {
    "id": "detail-6600",
    "role": "detail",
    "path": "generated/world-map-detail-6600.webp",
    "format": "webp",
    "width": 6600,
    "height": 6600,
    "metresPerPixel": 0.25,
    "bytes": 1407712,
    "sha256": "56448b8f8d2d15edf21c2fdd1a84487fa09759bae3f80132ad7e8d2e2d07709c",
    "quality": 60
  },
  {
    "id": "detail-3300",
    "role": "detail",
    "path": "generated/world-map-detail-3300.webp",
    "format": "webp",
    "width": 3300,
    "height": 3300,
    "metresPerPixel": 0.5,
    "bytes": 850166,
    "sha256": "2f77904656b31d3aa0a2c67e7b9a5b02a2f4b87c42ddcc2600f048597ef502dc",
    "quality": 82
  },
  {
    "id": "detail-1650",
    "role": "detail",
    "path": "generated/world-map-detail-1650.webp",
    "format": "webp",
    "width": 1650,
    "height": 1650,
    "metresPerPixel": 1,
    "bytes": 357490,
    "sha256": "fd0bd680c8a732fef416a03400c9767d5d4ecdfabdcbba620fce6a615b5887e9",
    "quality": 86
  }
] as const;